requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
Faker==16.3.0
idna==3.4
phone-gen==2.2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
urllib3==1.26.14
